# Conflicts:
#	apps/commons/src/test/java/ai/metaheuristic/commons/yaml/TestBundleParamsYaml.java
#	apps/metaheuristic/src/test/java/ai/metaheuristic/ai/function/TestFunctionConfigSchemeValidation.java

+++ apps/metaheuristic/src/main/java/ai/metaheuristic/ai/dispatcher/commons/ArtifactCleanerAtDispatcher.java
@@ -19,15 +19,15 @@ import ai.metaheuristic.ai.Consts;
 import ai.metaheuristic.ai.Globals;
 import ai.metaheuristic.ai.dispatcher.batch.BatchCache;
 import ai.metaheuristic.ai.dispatcher.batch.BatchTopLevelService;
-import ai.metaheuristic.ai.dispatcher.beans.Batch;
+import ai.metaheuristic.ai.dispatcher.beans.*;
 import ai.metaheuristic.ai.dispatcher.cache.CacheService;
 import ai.metaheuristic.ai.dispatcher.event.DispatcherEventService;
 import ai.metaheuristic.ai.dispatcher.exec_context.ExecContextCache;
 import ai.metaheuristic.ai.dispatcher.exec_context.ExecContextService;
-import ai.metaheuristic.ai.dispatcher.exec_context.ExecContextSyncService;
 import ai.metaheuristic.ai.dispatcher.exec_context.ExecContextTopLevelService;
 import ai.metaheuristic.ai.dispatcher.repositories.*;
 import ai.metaheuristic.ai.dispatcher.source_code.SourceCodeCache;
+import ai.metaheuristic.ai.dispatcher.task.TaskQueueService;
 import ai.metaheuristic.ai.dispatcher.task.TaskTransactionalService;
 import ai.metaheuristic.ai.dispatcher.variable.VariableService;
 import ai.metaheuristic.ai.utils.CollectionUtils;
@@ -41,6 +41,7 @@ import java.time.LocalDate;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
 @SuppressWarnings("DuplicatedCode")
@@ -50,9 +51,12 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class ArtifactCleanerAtDispatcher {
 
-    public final Globals globals;
+    private final Globals globals;
     private final ExecContextTopLevelService execContextTopLevelService;
     private final ExecContextRepository execContextRepository;
+    private final ExecContextGraphRepository execContextGraphRepository;
+    private final ExecContextTaskStateRepository execContextTaskStateRepository;
+    private final ExecContextVariableStateRepository execContextVariableStateRepository;
     private final SourceCodeCache sourceCodeCache;
     private final BatchRepository batchRepository;
     private final CompanyRepository companyRepository;
@@ -68,9 +72,30 @@ public class ArtifactCleanerAtDispatcher {
     private final CacheService cacheService;
     private final ExecContextService execContextService;
     private final DispatcherEventRepository dispatcherEventRepository;
+    private final FunctionDataRepository functionDataRepository;
+
+    private static final AtomicInteger busy = new AtomicInteger(0);
+
+    private static boolean isBusy() {
+        return busy.get() > 0;
+    }
+
+    public static void setBusy() {
+        ArtifactCleanerAtDispatcher.busy.incrementAndGet();
+    }
+
+    public static void notBusy() {
+        int curr = ArtifactCleanerAtDispatcher.busy.decrementAndGet();
+        if (curr<0) {
+            throw new IllegalStateException("(curr<0)");
+        }
+    }
 
     public void fixedDelay() {
         TxUtils.checkTxNotExists();
+        if (isBusy() || !TaskQueueService.isQueueEmptyWithSync()) {
+            return;
+        }
 
         // do not change the order of calling
         deleteOrphanAndObsoletedBatches();
@@ -79,6 +104,20 @@ public class ArtifactCleanerAtDispatcher {
         deleteOrphanVariables();
         deleteOrphanCacheData();
         deleteObsoleteEvents();
+        deleteObsoleteFunctionData();
+    }
+
+    private void deleteObsoleteFunctionData() {
+        List<String> functionCodesInData = functionDataRepository.findAllFunctionCodes();
+        List<String> functionCodes = functionRepository.findAllFunctionCodes();
+        for (String functionCode : functionCodesInData) {
+            if (!functionCodes.contains(functionCode)) {
+                if (isBusy()) {
+                    return;
+                }
+                functionDataRepository.deleteByFunctionCode(functionCode);
+            }
+        }
     }
 
     private void deleteObsoleteEvents() {
@@ -94,13 +133,21 @@ public class ArtifactCleanerAtDispatcher {
         int period = DispatcherEventService.getPeriod(keepStartDate);
 
         List<Long> periodsForDelete = dispatcherEventRepository.getPeriodIdsBefore(period);
-        List<List<Long>> pages = CollectionUtils.parseAsPages(periodsForDelete, 10);
+        List<List<Long>> pages = CollectionUtils.parseAsPages(periodsForDelete, 20);
         for (List<Long> page : pages) {
+            if (isBusy()) {
+                return;
+            }
             if (page.isEmpty()) {
                 continue;
             }
-            log.info("Found obsolete events #{}", page);
-            dispatcherEventRepository.deleteAllByIdIn(page);
+            log.info("#510.140 Found obsolete events #{}", page);
+            try {
+                dispatcherEventRepository.deleteAllByIdIn(page);
+            }
+            catch (Throwable th) {
+                log.error("#510.200 dispatcherEventRepository.deleteAllByIdIn("+page+")", th);
+            }
         }
     }
 
@@ -110,6 +157,9 @@ public class ArtifactCleanerAtDispatcher {
         Set<Long> forDeletion = new HashSet<>();
         List<Object[]> batches = batchRepository.findAllIdAndCreatedOnAndDeleted();
         for (Object[] obj : batches) {
+            if (isBusy()) {
+                return;
+            }
             long batchId = ((Number)obj[0]).longValue();
             long createdOn = ((Number)obj[1]).longValue();
             boolean deleted = Boolean.TRUE.equals(obj[2]);
@@ -131,16 +181,104 @@ public class ArtifactCleanerAtDispatcher {
     }
 
     private void deleteOrphanExecContexts() {
+        List<Long> execContextIds = execContextRepository.findAllIds();
+        deleteOrphanExecContexts(execContextIds);
+        if (isBusy()) {
+            return;
+        }
+        deleteOrphanExecContextGraph(execContextIds);
+        if (isBusy()) {
+            return;
+        }
+        deleteOrphanExecContextTaskState(execContextIds);
+        if (isBusy()) {
+            return;
+        }
+        deleteOrphanExecContextVariableState(execContextIds);
+    }
+
+    private void deleteOrphanExecContexts(List<Long> execContextIds) {
         Set<Long> forDeletion = new HashSet<>();
-        List<Object[]> objs = execContextRepository.findAllExecContextIdWithSourceCodeId();
-        for (Object[] obj : objs) {
-            long sourceCodeId = ((Number)obj[1]).longValue();
-            if (sourceCodeCache.findById(sourceCodeId)==null) {
-                long execContextId = ((Number)obj[0]).longValue();
+        for (Long execContextId : execContextIds) {
+            ExecContextImpl ec = execContextCache.findById(execContextId);
+            if (ec==null) {
+                continue;
+            }
+            if (sourceCodeCache.findById(ec.sourceCodeId)==null ||
+                    (ec.rootExecContextId!=null && execContextCache.findById(ec.rootExecContextId)==null)) {
                 forDeletion.add(execContextId);
             }
         }
         execContextTopLevelService.deleteOrphanExecContexts(forDeletion);
+    }
+
+    private void deleteOrphanExecContextGraph(List<Long> execContextIds) {
+        Set<Long> execContextGraphIds = new HashSet<>();
+        for (Long execContextId : execContextIds) {
+            ExecContextImpl ec = execContextCache.findById(execContextId);
+            if (ec==null) {
+                continue;
+            }
+            execContextGraphIds.add(ec.execContextGraphId);
+        }
+        List<Long> allExecContextGraphIds = execContextGraphRepository.findAllIds();
+        for (Long allExecContextGraphId : allExecContextGraphIds) {
+            if (!execContextGraphIds.contains(allExecContextGraphId)) {
+                ExecContextGraph execContextGraph = execContextGraphRepository.findById(allExecContextGraphId).orElse(null);
+                if (execContextGraph==null || execContextGraph.createdOn==null ||
+                        execContextGraph.createdOn==0 || (System.currentTimeMillis()-execContextGraph.createdOn) < 3_600_000 ) {
+                    continue;
+                }
+                log.info("#510.240 Found orphan ExecContextGraph #{}", allExecContextGraphId);
+                execContextService.deleteOrphanExecContextGraph(allExecContextGraphId);
+            }
+        }
+    }
+
+    private void deleteOrphanExecContextTaskState(List<Long> execContextIds) {
+        Set<Long> execContextTaskStateIds = new HashSet<>();
+        for (Long execContextId : execContextIds) {
+            ExecContextImpl ec = execContextCache.findById(execContextId);
+            if (ec==null) {
+                continue;
+            }
+            execContextTaskStateIds.add(ec.execContextTaskStateId);
+        }
+        List<Long> allExecContextTaskStateIds = execContextTaskStateRepository.findAllIds();
+        for (Long allExecContextTaskStateId : allExecContextTaskStateIds) {
+            if (!execContextTaskStateIds.contains(allExecContextTaskStateId)) {
+                ExecContextTaskState execContextTaskState = execContextTaskStateRepository.findById(allExecContextTaskStateId).orElse(null);
+                if (execContextTaskState==null || execContextTaskState.createdOn==null ||
+                        execContextTaskState.createdOn==0 || (System.currentTimeMillis()-execContextTaskState.createdOn) < 3_600_000 ) {
+                    continue;
+                }
+                log.info("#510.280 Found orphan ExecContextTaskState #{}", allExecContextTaskStateId);
+                execContextService.deleteOrphanExecContextTaskState(allExecContextTaskStateId);
+            }
+        }
+    }
+
+    private void deleteOrphanExecContextVariableState(List<Long> execContextIds) {
+        Set<Long> execContextVariableStateIds = new HashSet<>();
+        for (Long execContextId : execContextIds) {
+            ExecContextImpl ec = execContextCache.findById(execContextId);
+            if (ec==null) {
+                continue;
+            }
+            execContextVariableStateIds.add(ec.execContextVariableStateId);
+        }
+        List<Long> allExecContextVariableStateIds = execContextVariableStateRepository.findAllIds();
+        for (Long allExecContextVariableStateId : allExecContextVariableStateIds) {
+            if (!execContextVariableStateIds.contains(allExecContextVariableStateId)) {
+                ExecContextVariableState execContextVariableState = execContextVariableStateRepository.findById(allExecContextVariableStateId).orElse(null);
+                if (execContextVariableState==null || execContextVariableState.createdOn==null ||
+                        execContextVariableState.createdOn==0 || (System.currentTimeMillis()-execContextVariableState.createdOn) < 3_600_000 ) {
+                    continue;
+                }
+                log.info("#510.320 Found orphan ExecContextVariableState #{}", allExecContextVariableStateId);
+                execContextService.deleteOrphanExecContextVariableState(allExecContextVariableStateId);
+            }
+        }
     }
 
     private void deleteOrphanTasks() {
@@ -149,10 +287,14 @@ public class ArtifactCleanerAtDispatcher {
         List<Long> taskExecContextIds = taskRepository.getAllExecContextIds();
         List<Long> execContextIds = execContextRepository.findAllIds();
 
+        //noinspection SimplifyStreamApiCallChains
         List<Long> orphanExecContextIds = taskExecContextIds.stream()
                 .filter(o->!execContextIds.contains(o)).collect(Collectors.toList());
 
         for (Long execContextId : orphanExecContextIds) {
+            if (isBusy()) {
+                return;
+            }
             if (execContextCache.findById(execContextId)!=null) {
                 log.warn("execContextId #{} still here", execContextId);
                 Long id = execContextRepository.findIdById(execContextId);
@@ -160,15 +302,28 @@ public class ArtifactCleanerAtDispatcher {
                     continue;
                 }
                 log.warn("execContextId #{} was deleted in db, clean up the cache", execContextId);
-                execContextService.deleteExecContextFromCache(execContextId);
+                try {
+                    execContextService.deleteExecContextFromCache(execContextId);
+                }
+                catch (Throwable th) {
+                    log.error("execContextService.deleteExecContextFromCache("+execContextId+")", th);
+                }
             }
 
             List<Long> ids;
             while (!(ids = taskRepository.findAllByExecContextId(Consts.PAGE_REQUEST_100_REC, execContextId)).isEmpty()) {
-                List<List<Long>> pages = CollectionUtils.parseAsPages(ids, 10);
+                List<List<Long>> pages = CollectionUtils.parseAsPages(ids, 20);
                 for (List<Long> page : pages) {
+                    if (page.isEmpty() || isBusy()) {
+                        return;
+                    }
                     log.info("Found orphan task, execContextId: #{}, tasks #{}", execContextId, page);
-                    ExecContextSyncService.getWithSyncNullable(execContextId, () -> taskTransactionalService.deleteOrphanTasks(page));
+                    try {
+                        taskTransactionalService.deleteOrphanTasks(page);
+                    }
+                    catch (Throwable th) {
+                        log.error("taskTransactionalService.deleteOrphanTasks("+execContextId+")", th);
+                    }
                 }
             }
         }
@@ -178,6 +333,7 @@ public class ArtifactCleanerAtDispatcher {
         List<Long> variableExecContextIds = variableRepository.getAllExecContextIds();
         List<Long> execContextIds = execContextRepository.findAllIds();
 
+        //noinspection SimplifyStreamApiCallChains
         List<Long> orphanExecContextIds = variableExecContextIds.stream()
                 .filter(o->!execContextIds.contains(o)).collect(Collectors.toList());
 
@@ -189,13 +345,21 @@ public class ArtifactCleanerAtDispatcher {
 
             List<Long> ids;
             while (!(ids = variableRepository.findAllByExecContextId(Consts.PAGE_REQUEST_100_REC, execContextId)).isEmpty()) {
-                List<List<Long>> pages = CollectionUtils.parseAsPages(ids, 5);
+                List<List<Long>> pages = CollectionUtils.parseAsPages(ids, 20);
                 for (List<Long> page : pages) {
+                    if (isBusy()) {
+                        return;
+                    }
                     if (page.isEmpty()) {
                         continue;
                     }
                     log.info("Found orphan variables, execContextId: #{}, variables #{}", execContextId, page);
-                    ExecContextSyncService.getWithSyncNullable(execContextId, () -> variableService.deleteOrphanVariables(page));
+                    try {
+                        variableService.deleteOrphanVariables(page);
+                    }
+                    catch (Throwable th) {
+                        log.error("variableService.deleteOrphanVariables("+execContextId+")", th);
+                    }
                 }
             }
         }
@@ -205,20 +369,34 @@ public class ArtifactCleanerAtDispatcher {
         List<String> funcCodes = functionRepository.findAllFunctionCodes();
         Set<String> currFuncCodes = cacheProcessRepository.findAllFunctionCodes();
 
+        //noinspection SimplifyStreamApiCallChains
         List<String> missingCodes = currFuncCodes.stream().filter(currFuncCode -> !funcCodes.contains(currFuncCode)).collect(Collectors.toList());
 
         for (String funcCode : missingCodes) {
             List<Long> ids;
             while (!(ids = cacheProcessRepository.findByFunctionCode(Consts.PAGE_REQUEST_100_REC, funcCode)).isEmpty()) {
-                List<List<Long>> pages = CollectionUtils.parseAsPages(ids, 5);
+                List<List<Long>> pages = CollectionUtils.parseAsPages(ids, 20);
                 for (List<Long> page : pages) {
+                    if (isBusy()) {
+                        return;
+                    }
                     if (page.isEmpty()) {
                         continue;
                     }
                     log.info("Found orphan cache entries, funcCode: #{}, cacheProcessIds #{}", funcCode, page);
-                    cacheService.deleteCacheProcesses(page);
+                    try {
+                        cacheService.deleteCacheProcesses(page);
+                    }
+                    catch (Throwable th) {
+                        log.error("cacheService.deleteCacheProcesses("+page+")", th);
+                    }
                     for (Long cacheProcessId : page) {
-                        cacheService.deleteCacheVariable(cacheProcessId);
+                        try {
+                            cacheService.deleteCacheVariable(cacheProcessId);
+                        }
+                        catch (Throwable th) {
+                            log.error("cacheService.deleteCacheVariable("+page+")", th);
+                        }
                     }
                 }
             }
